+++ tools/betterBestArrow/tool.js
@@ -8,7 +8,7 @@
         name: 'betterBestArrow',
         category: 'analysis2',
         type: 'multiple',
-        possibleValues: ['enabled','localEval','allMoves'],
+        possibleValues: ['enabled','localEval','allMoves','justOrientation'],
         defaultValue: 'enabled,localEval,allMoves',
         advanced: true
       }
@@ -20,14 +20,16 @@
         'options.betterBestArrow': 'Improved best move arrow',
         'betterBestArrow.enabled': 'Enabled',
         'betterBestArrow.localEval': 'Use local evaluation',
-        'betterBestArrow.allMoves': 'On all moves'
+        'betterBestArrow.allMoves': 'On all moves',
+        'betterBestArrow.justOrientation': 'Only moves for orientation side'
       },
       'ro-RO': {
         'options.analysis2': 'Analiz\u0103 - m\u0103run\u0163i\u015furi',
         'options.betterBestArrow': 'S\u0103geat\u0103 cu cea mai bun\u0103 mutare \u00eembun\u0103t\u0103\u0163it\u0103',
         'betterBestArrow.enabled': 'Activat',
         'betterBestArrow.localEval': 'Folose\u015fte evaluarea local\u0103',
-        'betterBestArrow.allMoves': 'La toate mut\u0103rile'
+        'betterBestArrow.allMoves': 'La toate mut\u0103rile',
+        'betterBestArrow.justOrientation': 'Doar mut\u0103ri ale p\u0103r\u0163ii orient\u0103rii'
       }
     }
 
@@ -38,7 +40,8 @@
       this.options = {
         enabled: lt.isOptionSet(value, 'enabled'),
         localEval: lt.isOptionSet(value, 'localEval'),
-        allMoves: lt.isOptionSet(value, 'allMoves')
+        allMoves: lt.isOptionSet(value, 'allMoves'),
+        justOrientation: lt.isOptionSet(value, 'justOrientation')
       };
 
       const lichess = lt.lichess;
@@ -49,6 +52,7 @@
       $('body').toggleClassSafe('lichessTools-betterBestArrow',!!value);
       analysis.setAutoShapes = lt.unwrapFunction(analysis.setAutoShapes,'betterBestArrow');
       if (!this.options.enabled) return;
+      let currBest;
       analysis.setAutoShapes = lt.wrapFunction(analysis.setAutoShapes,{
         id: 'betterBestArrow',
         before: ($this, ...args) => {
@@ -63,10 +67,18 @@
           const best = ebest && cbest && node.ceval?.depth < 15
             ? ebest
             : cbest || ebest;
-          if (best && node.eval?.best != best) {
-            node.eval ||= {};
-            node.eval.best = best;
+          if (!this.options.justOrientation || analysis.getOrientation() != analysis.turnColor()) {
+            if (best && node.eval?.best != best) {
+              node.eval ||= {};
+              node.eval.best = best;
+            }
           }
+          currBest = best;
+        },
+        after: ($this, result, ...args)=>{
+          if (!this.options.justOrientation) return;
+          if (analysis.getOrientation()!=analysis.turnColor()) return;
+          lt.arrayRemoveAll(analysis.chessground?.state?.drawable?.autoShapes,s=>s.brush=='paleGreen' && s.orig+s.dest==currBest)
         }
       });
     }