+++ HISTORY.md
@@ -4,6 +4,7 @@ History of features added to LiChess Tools in time (reversed order)
 
 - Time controls theme
 - FEN based wiki tooltip on study opening link
+- option to show best move just for orientation side
 
 06 Jun